[UPDATE]Se crea el Carousel de imagenes ocn imagenes aleatorias, se ánade el viewmodel al fragment y luego se añade el fragment al activity
