+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,4 @@
 <resources>
     <string name="app_name">ONGAndroid</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="lastest_news">Ultimas Novedades</string>
 </resources>
